+++ lib/event.rb
@@ -1,0 +1,6 @@
+class Event
+
+  def initialize(name)
+    @name = name
+  end
+end
