+++ spec/event_spec.rb
@@ -1,0 +1,20 @@
+require './lib/item'
+require './lib/food_truck'
+require './lib/event'
+
+RSpec.describe Event do
+  before :each do
+    @event = Event.new("South Pearl Street Farmers Market")
+
+    @food_truck1 = FoodTruck.new("Rocky Mountain Pies")
+
+    @item1 = Item.new({name: 'Peach Pie (Slice)', price: "$3.75"})
+    @item2 = Item.new({name: 'Apple Pie (Slice)', price: '$2.50'})
+    @item3 = Item.new({name: "Peach-Raspberry Nice Cream", price: "$5.30"})
+    @item4 = Item.new({name: "Banana Nice Cream", price: "$4.25"})
+  end
+
+  it 'exists' do
+    expect(@event).to be_an_instance_of(Event)
+  end
+end
